# api/openapi.yaml
openapi: 3.1.0
info:
  title: CloudBank API
  version: 0.1.0
  description: >
    HTTP API for CloudBank, a self-hosted web port of HomeBank. This document is
    the single source of truth for the contract: the Go server interfaces and
    the TypeScript client types are both generated from it. Resources are added
    milestone by milestone; this initial version covers the health and ping
    endpoints established by the server skeleton.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
paths:
...
  /api/v1/setup:
    post:
      operationId: postSetup
      summary: Create the first (admin) user
      description: Allowed only when no user exists yet; creates an administrator and starts a session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Admin created and logged in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Setup has already been completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Logged in, or a two-factor code is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429":
          description: Too many attempts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/auth/me:
    get:
      operationId: getMe
      summary: Current authenticated user
      responses:
        "200":
          description: The current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateMe
      summary: Update the current user's language, theme and UI preferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMeRequest" }
      responses:
        "200": { description: The updated user., content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/wallets/{walletId}/bank/enablebanking/auth:
    parameters:
      - { name: walletId, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      operationId: startEnableBankingAuth
      summary: Begin a bank authorization; returns the URL to redirect the user to their bank
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, properties: { aspspName: { type: string }, aspspCountry: { type: string }, name: { type: string }, redirectUrl: { type: string } }, required: [aspspName, aspspCountry, redirectUrl] }
      responses:
        "200": { description: The authorization URL and state., content: { application/json: { schema: { type: object, properties: { url: { type: string }, state: { type: string } }, required: [url, state] } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/v1/admin/users:
    get:
      operationId: listUsers
      summary: List users (admin)
      responses:
        "200":
          description: All users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createUser
      summary: Create a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Username already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/users/{id}/disable:
    post:
      operationId: setUserDisabled
      summary: Enable or disable a user (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disabled: { type: boolean }
              required: [disabled]
      responses:
        "204": { description: Updated. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{id}/password:
    post:
      operationId: resetUserPassword
      summary: Reset a user's password (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, minLength: 8 }
              required: [password]
      responses:
        "204": { description: Password reset. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Administrator access required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
    User:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        email: { type: string }
        isAdmin: { type: boolean }
        locale: { type: string }
        theme: { type: string }
        preferences: { type: object, additionalProperties: true, description: "opaque UI preferences blob" }
        disabled: { type: boolean }
        twoFactorEnabled: { type: boolean }
        createdAt: { type: string }
      required: [id, username, email, isAdmin, locale, theme, disabled, createdAt]
    UpdateMeRequest:
      type: object
      properties:
        locale: { type: string }
        theme: { type: string }
        preferences: { type: object, additionalProperties: true }
...
    Credentials:
      type: object
      properties:
        username: { type: string }
        email: { type: string }
        password: { type: string, minLength: 8 }
        totpCode: { type: string, description: "second factor (TOTP or a recovery code); required when the account has 2FA and login returned totpRequired" }
      required: [username, password]
    LoginResult:
      description: "On success the logged-in User; when 2FA is on and no code was given, { totpRequired: true } (the password was accepted, resend with totpCode)."
      oneOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            totpRequired: { type: boolean, enum: [true] }
          required: [totpRequired]
    CreateUserRequest:
      type: object
      properties:
        username: { type: string }
        email: { type: string }
        password: { type: string, minLength: 8 }
        isAdmin: { type: boolean }
      required: [username, password]
...